add app state props to components
