+++ src/actions/actionCreator.js
@@ -1,13 +1,13 @@
 import axios from 'axios';
-import { useHistory } from 'react-router-dom';
 import * as actions from './actionTypes';
 
-const history = useHistory();
+// REQUEST ACTIONS
 // const requestingData = () => ({ type: actions.REQUESTING_DATA });
 // const requestingFailed = () => ({ type: actions.REQUESTING_FAILED });
 const sendingData = () => ({ type: actions.SENDING_DATA });
 const sendingFailed = () => ({ type: actions.SENDING_FAILED });
 
+// SIGNUP ACTIONS
 export const receivedNewUserData = (response) => ({
   type: actions.SIGN_UP,
   payload: response.data,
@@ -17,7 +17,7 @@ export const addNewUser = (name, email, password) => async (dispatch) => {
   try {
     dispatch(sendingData());
     const response = await axios.post(
-      'http://localhost:3000/api/v1/users',
+      'http://localhost:3001/api/v1/users',
       {
         name,
         email,
@@ -26,10 +26,11 @@ export const addNewUser = (name, email, password) => async (dispatch) => {
       { withCredentials: true },
     );
     if (response.data.status === 'created') {
-      history.push('/');
+      console.log('user is', response.data);
       dispatch(receivedNewUserData(response));
     }
-  } catch (e) {
+  } catch (error) {
+    console.log('Regsitration Errors', error.message);
     dispatch(sendingFailed());
   }
 };
