+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import App from './components/App';
+import App from './containers/App';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
 import Store from './Store';
